make action_timetableUserId run with book.js
book.js needs a userId. this was the studentId, because only students can booking a slot. but teacher can set ("booking") a slot
to a student. So the userId is sometimes a studentId or a teacherId. it depends who is logged in. students see teacher's
timetable and the other way round